+++ services/settlement-service/src/main/java/com/northstar/settlement/persistence/SettlementRow.java
@@ -1,11 +1,8 @@
 package com.northstar.settlement.persistence;
 
-import java.math.BigDecimal;
-
 /** One SETTLEMENT row, column for column. */
 public record SettlementRow(int settlementId, int claimId,
-        BigDecimal coveredAmount, BigDecimal deductibleApplied,
-        BigDecimal depreciation, boolean cappedAtLimit,
-        BigDecimal settlementAmount, String calculatedBy,
+        double coveredAmount, double deductibleApplied, double depreciation,
+        boolean cappedAtLimit, double settlementAmount, String calculatedBy,
         String calculatedDate) {
 }
